feat: (maybe?) simplify ssh shit
Sort of. It would be nice to not have to generate user keys. On a single
user system, such as ever system I currently own and run, I don't really
need to differentiate between the host and me. The host key should be
sufficient. I just need to tell sshd to add the host key?